+++ src/Utility/ThumbManager.php
@@ -25,6 +25,12 @@ class ThumbManager
     use ThumbTrait;
 
     /**
+     * Supported formats
+     * @var array
+     */
+    public static $supportedFormats = ['bmp', 'gif', 'ico', 'jpg', 'png', 'psd', 'tiff'];
+
+    /**
      * Internal method to clear thumbnails
      * @param array $filenames Filenames
      * @return int|bool Number of thumbnails deleted otherwise `false` in case of error
@@ -52,7 +58,7 @@ class ThumbManager
      */
     protected function _find($regexpPattern = null, $sort = false)
     {
-        $regexpPattern = $regexpPattern ?: sprintf('[a-z0-9]{32}_[a-z0-9]{32}\.(%s)', implode('|', $this->getSupportedFormats()));
+        $regexpPattern = $regexpPattern ?: sprintf('[a-z0-9]{32}_[a-z0-9]{32}\.(%s)', implode('|', self::$supportedFormats));
 
         return (new Folder($this->getPath()))->find($regexpPattern, $sort);
     }
@@ -89,7 +95,7 @@ class ThumbManager
      */
     public function get($path, $sort = false)
     {
-        $regexpPattern = sprintf('%s_[a-z0-9]{32}\.(%s)', md5($this->resolveFilePath($path)), implode('|', $this->getSupportedFormats()));
+        $regexpPattern = sprintf('%s_[a-z0-9]{32}\.(%s)', md5($this->resolveFilePath($path)), implode('|', self::$supportedFormats));
 
         return $this->_find($regexpPattern, $sort);
     }